+++ CHANGELOG.md
@@ -2,6 +2,38 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 2.9.0 - TBD
+
+### Added
+
+- [#46](https://github.com/zendframework/zend-log/pull/46) adds the ability to
+  specify log writer, formatter, filter, and processor plugin configuration via
+  the new top-level keys:
+  - `log_filters`
+  - `log_formatters`
+  - `log_processors`
+  - `log_writers`
+  These follow the same configuration patterns as any other service
+  manager/plugin manager as implemented by zend-servicemanager.
+
+  Additionally, you can now specify filer, formatter, and processor *services*
+  when specifying writer configuration for a logger, as these are now backed
+  by the above plugin managers.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- [#38](https://github.com/zendframework/zend-log/pull/38) adds the `MongoDb`
+  writer to the list of available writer plugins; the writer was added in a
+  previous release, but never enabled within the default set of writers.
+
 ## 2.8.4 - TBD
 
 ### Added